add : SearchBar 추가 구현 Edit : interface 추가
